Run work
Added some logic to deal with "Run" event.
Swapped the "Main" in "Brandon" pckg to "TestView" in tests package.
Swapped css to controller.
Deleted Brandon pckg as it's no longer needed

Still need:
Event handling logic, especially for randomization.
Fix the display of the gantt and process info charts
Add the averages location on the chart which appears when the gantt calculates.